requirements.txt
discord.py
typing
requests